+++ R/SoR_SE.R
@@ -43,7 +43,7 @@
 #'        fitting.method = c("Exponential", "Gamma")[1])
 #'
 SoR.SE <- function (data, const = 0, threshold = c("mean", "const")[1],
-                    se.method = c("IFiid","IFcor","IFcorAdapt","IFcorPW","BOOTiid","BOOTcor")[c(1,3)],
+                    se.method = c("IFiid","IFcor","IFcorAdapt","IFcorPW","BOOTiid","BOOTcor")[c(1,4)],
                     cleanOutliers = FALSE, fitting.method = c("Exponential", "Gamma")[1], d.GLM.EN = 5,
                     freq.include = c("All", "Decimate", "Truncate")[1], freq.par = 0.5,
                     corOut = c("none", "retCor","retIFCor")[1],